Changed money format in XML file

+++ FinancialData.h
@@ -2,6 +2,7 @@
 #define FINANCIALDATA_H_INCLUDED
 
 #include <iostream>
+#include <iomanip>
 
 using namespace std;
 
@@ -10,18 +11,18 @@ class FinancialData
     int id;
     int userId;
     string category;
-    int moneyAmmount;
+    float moneyAmmount;
     int date;
 public:
     void setId(int newId);
     void setUserId(int newUserId);
     void setCategory(string newCategory);
-    void setMoneyAmmount(int newMoneyAmmount);
+    void setMoneyAmmount(float newMoneyAmmount);
     void setDate(int newDate);
     int getId();
     int getUserId();
     string getCategory();
-    int getMoneyAmmount();
+    float getMoneyAmmount();
     int getDate();
 };
